wip: created koboSyncScheduler
